+++ docker/download_pods.clj
@@ -4,5 +4,7 @@
 
 (def OPTIONS (edn/read-string (slurp (io/file "/project/bb.edn"))))
 
-(doseq [[s v] (:pods (:runtime (:holy-lambda/options OPTIONS)))]
+(doseq [[s v] (:pods (or (:runtime (:holy-lambda/options OPTIONS))
+                         (:backend (:holy-lambda/options OPTIONS))))]
+  (println "[holy-lambda] Downloading pod =" s "v =" v)
   (pods/load-pod s v))