+++ modules/holy-lambda-babashka-tasks/src/holy_lambda/tasks.clj
@@ -5,17 +5,71 @@
    [clojure.java.shell :as csh]
    [clojure.edn :as edn]
    [cheshire.core :as json]
-   [babashka.deps :as deps]
-   [babashka.fs :as fs]
    [babashka.curl :as curl]
+   [babashka.fs :as fs]
    [babashka.process :as p]
    [clojure.java.io :as io]
    [holy-lambda.refl :as refl])
   (:refer-clojure :exclude [spit]))
 
-(def TASKS_VERSION "0.5.0")
+(def TASKS_VERSION "0.5.1")
 
-(deps/add-deps {:deps {'clojure-term-colors/clojure-term-colors {:mvn/version "0.1.0"}}})
+;; Coloring taken from https://github.com/trhura/clojure-term-colors/blob/master/src/clojure/term/colors.clj
+;; Included as is to prevent the syncing of deps
+(defn- escape-code
+  [i]
+  (str "\033[" i "m"))
+
+(def ^:dynamic *colors*
+  "foreground color map"
+  (zipmap [:grey :red :green :yellow
+           :blue :magenta :cyan :white]
+          (map escape-code
+               (range 30 38))))
+
+(def ^:dynamic *highlights*
+  "background color map"
+  (zipmap [:on-grey :on-red :on-green :on-yellow
+           :on-blue :on-magenta :on-cyan :on-white]
+          (map escape-code
+            (range 40 48))))
+
+(def ^:dynamic *attributes*
+  "attributes color map"
+  (into {}
+        (filter (comp not nil? key)
+                (zipmap [:bold, :dark, nil, :underline,
+                         :blink, nil, :reverse-color, :concealed]
+                        (map escape-code (range 1 9))))))
+
+(def ^:dynamic *reset* (escape-code 0))
+
+;; Bind to true to have the colorize functions not apply coloring to
+;; their arguments.
+(def ^:dynamic *disable-colors* nil)
+
+(defmacro define-color-function
+  "define a function `fname' which wraps its arguments with
+        corresponding `color' codes"
+  [fname color]
+  (let [fname (symbol (name fname))
+        args (symbol 'args)]
+    `(defn ~fname [& ~args]
+       (if-not *disable-colors*
+         (str (clojure.string/join (map #(str ~color %) ~args)) ~*reset*)
+         (apply str ~args)))))
+
+(defn define-color-functions-from-map
+  "define functions from color maps."
+  [colormap]
+  (eval `(do ~@(map (fn [[color escape-code]]
+                `(println ~color ~escape-code)
+                `(define-color-function ~color ~escape-code))
+                    colormap))))
+
+(define-color-functions-from-map *colors*)
+(define-color-functions-from-map *highlights*)
+(define-color-functions-from-map *attributes*)
 
 (defn env-true?
   [env]
@@ -23,9 +77,6 @@
     (contains? #{"true" "1"} prop)))
 
 (def HL_DEBUG? (env-true? "HL_DEBUG"))
-
-(require
- '[clojure.term.colors :refer [underline blue yellow red green]])
 
 ;;;; [START] HELPERS
 
@@ -220,17 +271,13 @@
 
 (def CLJ_ALIAS_KEY (or (some-> (System/getenv "HL_CLJ_ALIAS") string->keyword)
                        (some-> (:clj-alias BUILD) string->keyword)))
+(when CLJ_ALIAS_KEY
+  (hpr (pre (str "HL_CLJ_ALIAS and :clj-alias are deprecated. Use HL_COMPILE_CMD or :compile-cmd instead"))))
 
-(when (and CLJ_ALIAS_KEY (not (keyword? CLJ_ALIAS_KEY)))
-  (hpr (pre "Defined") (accent "build:clj-alias") (pre "should be a keyword")))
+(def COMPILE_CMD (or (System/getenv "HL_COMPILE_CMD")
+                     (:compile-cmd BUILD)))
 
-(when (and CLJ_ALIAS_KEY (not (CLJ_ALIAS_KEY (:aliases (read-string (slurp "deps.edn"))))))
-  (hpr (pre "Defined")
-       (accent "build:clj-alias")
-       (accent CLJ_ALIAS_KEY)
-       (pre "does not exists in") (accent "deps.edn")))
-
-(def CLJ_ALIAS (some-> CLJ_ALIAS_KEY name))
+(def COMPILE_CMD_ERROR (pre "Neither HL_COMPILE_CMD nor :build:compile-cmd has been specified!"))
 
 (defn stat-file
   [filename]
@@ -283,7 +330,8 @@
           DOCKER_VOLUMES_CONF)))
 
 (def DOCKER_NETWORK (:network DOCKER))
-(def RUNTIME (:runtime OPTIONS))
+(def RUNTIME (or (:runtime OPTIONS)
+                 (:backend OPTIONS)))
 (def NATIVE_IMAGE_ARGS (if-let [args (seq (:native-image-args RUNTIME))]
                          (s/join " " args)
                          nil))
@@ -311,10 +359,8 @@
                       {}))
       (s/trim (:out result)))))
 
-(def ENTRYPOINT (:entrypoint (:runtime OPTIONS)))
 (def OUTPUT_JAR_PATH ".holy-lambda/build/output.jar")
 (def OUTPUT_JAR_PATH_WITH_AGENT ".holy-lambda/build/output-agent.jar")
-(def HOLY_LAMBDA_DEPS_PATH ".holy-lambda/clojure/deps.edn")
 (def REQUIRED_COMMANDS ["aws" "sam" "bb" "docker" "clojure" "zip" "id" "bash"])
 (def NATIVE_CONFIGURATIONS_PATH "resources/native-configuration")
 (def NATIVE_CONFIGURATIONS_RESOURCE_CONFIG_FILE_PATH "resources/native-configuration/resource-config.json")
@@ -337,13 +383,6 @@
       (catch Exception e
         (hpr (ex-message e))
         (System/exit 1)))))
-
-(defn exit-if-not-synced!
-  []
-  (when (and (not HL_NO_DOCKER?)
-             (not (file-exists? ".holy-lambda/clojure/deps.edn")))
-    (hpr (pre "Project has not been synced yet. Run") (accent "hl:sync") (pre "before running this command!"))
-    (System/exit 1)))
 
 (def USER_GID
   (str (s/trim (shs "id -u"))
@@ -401,63 +440,23 @@
 
 (defn deps-sync--babashka
   []
-  (when-not (empty? (:pods (:runtime OPTIONS)))
+  (shell "mkdir -p .holy-lambda/bb-clj-deps")
+  (shell "bb" "-cp" (s/trim (shs "bb" "--clojure" "-Spath")) "--uberjar" ".holy-lambda/bb-clj-deps/.m2/libs.jar")
+  (shell "bash -c \"cd .holy-lambda/bb-clj-deps/.m2 && unzip -oq libs.jar && rm -Rf libs.jar\"")
+  (when-not (empty? (:pods RUNTIME))
     (hpr "Babashka pods found! Syncing" (str (accent "babashka pods") ".") "Pods should be distributed via a layer which points to" (accent ".holy-lambda/pods"))
     (docker-run "cd / && download_pods")
     (when (file-exists? ".holy-lambda/.babashka")
       (shell "rm -Rf .holy-lambda/pods")
       (shell "mkdir -p .holy-lambda/pods")
-      (shell "cp -R .holy-lambda/.babashka .holy-lambda/pods/")))
-  (shell "bash -c \"mkdir -p .holy-lambda/bb-clj-deps && cp -R .holy-lambda/.m2 .holy-lambda/bb-clj-deps/\""))
+      (shell "cp -R .holy-lambda/.babashka .holy-lambda/pods/"))))
 
-(def tasks-deps-edn
-  {:mvn/local-repo ".holy-lambda/.m2"
-   :aliases {:uberjar {:replace-deps {'com.github.seancorfield/depstar {:mvn/version "2.0.216"}}
-                       :exec-fn 'hf.depstar/uberjar
-                       :exec-args {}}}})
-
-(defn deps-sync--deps
+(defn hl:babashka:sync
+  "     \033[0;31m>\033[0m Syncs dependencies bb.edn :backend:pods and dependencies to .holy-lambda directory. "
   []
-  (stat-file "deps.edn")
-  (hpr "Syncing project and holy-lambda" (accent "deps.edn"))
-  (docker-run (str "clojure -A:" (and CLJ_ALIAS (str CLJ_ALIAS ":")) "uberjar -P"))
-  (deps-sync--babashka))
-
-(defn hl:babashka-sync
-  "     \033[0;31m>\033[0m Syncs project & dependencies from either:
-       \t\t            - \033[0;31m<Clojure>\033[0m  deps.edn
-       \t\t            - \033[0;31m<Babashka>\033[0m bb.edn:runtime:pods"
-  []
-  (print-task "hl:babashka-sync")
-  (when (and (not HL_NO_DOCKER?)
-             (not (file-exists? ".holy-lambda/clojure/deps.edn")))
-    (hpr "Project not synced yet. Syncing with docker image!")
-    (let [cid (gensym "holy-lambda")]
-      (shs "docker" "create" "--user" USER_GID "-ti" "--name" (str cid) IMAGE_CORDS "bash")
-      (shs "docker" "cp" (str cid ":/project/.holy-lambda") ".")
-      (shs "docker" "rm" "-f" (str cid))))
-
-  (if HL_NO_DOCKER?
-    (shell "mkdir -p .holy-lambda")
-    ;; Correct holy-lambda deps.edn
-    (spit HOLY_LAMBDA_DEPS_PATH (edn->pp-sedn tasks-deps-edn)))
-
-  ;; Sync
-  (deps-sync--deps)
-
-  (when-not (file-exists? ".holy-lambda")
-    (hpr (pre "Unable to sync docker image content with") (accent ".holy-lambda") (pre "project directory!")))
-
-  (when-not (file-exists? ".holy-lambda/clojure")
-    (hpr (pre "Project did not sync properly. Remove .holy-lambda directory and run") (accent "hl:sync")))
-
-  (hpr "Sync completed!"))
-
-(defn stack-files-check--native
-  []
-  (when-not (file-exists? ".holy-lambda/build/latest.zip")
-    (hpr (pre "No") (accent ".holy-lambda/build/latest.zip") (pre "found! Run") (accent "native:executable"))
-    (System/exit 1)))
+  (print-task "hl:babashka:sync")
+  (deps-sync--babashka)
+  (hpr "Babashka deps & pods sync completed"))
 
 (defn stack-files-check--jar
   []
@@ -465,56 +464,39 @@
     (hpr (pre "No") (accent OUTPUT_JAR_PATH) (pre "found! Run") (accent "hl:compile"))
     (System/exit 1)))
 
-(defn stack-files-check
-  [& [check]]
-  (when-not (file-exists? ".holy-lambda")
-    (hpr (pre "No") (accent ".holy-lambda") (pre "directory! Run") (accent "hl:sync"))
-    (System/exit 1))
-
-  (case check
-    :native (do
-              (stack-files-check--jar)
-              (stack-files-check--native))
-    :clojure (stack-files-check--jar)
-    :babashka nil
-    nil))
-
 (defn build-stale?
   []
   (if-not (file-exists? "src")
-    true
+    false
     (boolean (seq (fs/modified-since OUTPUT_JAR_PATH (fs/glob "src" "**/**.{clj,cljc,cljs}"))))))
 
 (defn hl:native:conf
   "     \033[0;31m>\033[0m Provides native configurations for the application"
   []
   (print-task "hl:native:conf")
-  (exit-if-not-synced!)
-  (stack-files-check :default)
-
+  (stack-files-check--jar)
   (io/make-parents (str NATIVE_CONFIGURATIONS_PATH "/traces.json"))
-
+  (when (build-stale?)
+    (hpr (prw "Build is stale. Consider recompilation via") (accent "hl:compile")))
+  (when-not COMPILE_CMD (hpr COMPILE_CMD_ERROR))
+  (shell "bash -c \"rm -Rf .cpcache .holy-lambda/build/output-agent.jar && mv .holy-lambda/build/output.jar .holy-lambda/build/output-temp.jar\"")
   (hpr "Compiling with agent support!")
-  (shell "rm -Rf .cpcache .holy-lambda/build/output-agent.jar")
-  (docker-run (str "USE_AGENT_CONTEXT=true clojure -X:uberjar :aliases '" (str [CLJ_ALIAS_KEY]) "' :aot '[\"" (str ENTRYPOINT) "\"]' " ":jvm-opts '[\"-Dclojure.compiler.direct-linking=true\", \"-Dclojure.spec.skip-macros=true\"]' :jar " OUTPUT_JAR_PATH_WITH_AGENT " :main-class " (str ENTRYPOINT)))
-
+  (binding [HL_NO_DOCKER? true] (docker-run (str "USE_AGENT_CONTEXT=true " COMPILE_CMD)))
+  (shell "bash -c \"mv .holy-lambda/build/output.jar .holy-lambda/build/output-agent.jar && mv .holy-lambda/build/output-temp.jar .holy-lambda/build/output.jar\"")
   (hpr "Generating traces to ignore unnecessary reflection entries!")
   (docker-run (str JAVA_COMMAND
                    " -agentlib:native-image-agent="
                    "trace-output=" (str NATIVE_CONFIGURATIONS_PATH "/traces.json")
                    " "
                    "-Dexecutor=native-agent -jar " OUTPUT_JAR_PATH_WITH_AGENT))
-
   (hpr "Generating native-configurations!")
   (docker-run (str JAVA_COMMAND
                    " -agentlib:native-image-agent="
                    "config-merge-dir=" NATIVE_CONFIGURATIONS_PATH
                    " "
                    "-Dexecutor=native-agent -jar " OUTPUT_JAR_PATH_WITH_AGENT))
-
   (hpr "Cleaning up reflection-config.json!")
   (refl/clean-reflection-config!)
-
   (if-not (file-exists? NATIVE_CONFIGURATIONS_RESOURCE_CONFIG_FILE_PATH)
     (hpr (pre "Native configurations generation failed!"))
     (let [resource-config (json/parse-string (slurp (io/file NATIVE_CONFIGURATIONS_RESOURCE_CONFIG_FILE_PATH)))]
@@ -550,7 +532,6 @@ set -e
 \n----------------------------------------------------------------\n"
   [& args]
   (print-task "hl:native:executable")
-  (exit-if-not-synced!)
   (stack-files-check--jar)
 
   (when (build-stale?)
@@ -558,20 +539,16 @@ set -e
 
   (when-not (file-exists? NATIVE_CONFIGURATIONS_PATH)
     (hpr (prw "No native configurations has been generated. Native image build may fail. Run") (accent "native:conf") (prw "to generate native configurations.")))
-
-  ;; Copy then build
   (shell-no-exit "bash -c \"[ -d resources/native-configuration ] && cp -rf resources/native-configuration .holy-lambda/build/\"")
-
   (docker-run (str "cd .holy-lambda/build/ && " NATIVE_IMAGE_COMMAND " -jar output.jar -H:ConfigurationFileDirectories=native-configuration"
                    (when NATIVE_IMAGE_ARGS
                      (str " " NATIVE_IMAGE_ARGS))))
-
   (if-not (file-exists? ".holy-lambda/build/output")
     (hpr (pre "Native image failed to create executable. Fix your build! Skipping next steps"))
     (do
       (spit ".holy-lambda/build/bootstrap" (bootstrap-file))
       (when (and NATIVE_DEPS_PATH (file-exists? NATIVE_DEPS_PATH))
-        (hpr "Copying" (accent ":runtime:native-deps"))
+        (hpr "Copying" (accent ":backend:native-deps"))
         (shell (str "cp -R " NATIVE_DEPS_PATH " .holy-lambda/build/")))
       (hpr "Bundling artifacts...")
       (shell "bash -c \"cd .holy-lambda/build && chmod +x bootstrap\"")
@@ -583,24 +560,20 @@ set -e
 
 (defn hl:compile
   "     \033[0;31m>\033[0m Compiles sources if necessary
-  \t\t            - \033[0;31m:force\033[0m - force compilation even if sources did not change
-  \t\t            - \033[0;31m:local\033[0m - force compilation even if sources did not change"
+  \t\t            - \033[0;31m:force\033[0m - force compilation even if sources did not change"
   [& args]
   (print-task "hl:compile")
-  (exit-if-not-synced!)
-  (let [{:keys [force local]} (norm-args args)]
+  (let [{:keys [force]} (norm-args args)]
     (when (and (not (build-stale?)) (not force))
-        (hpr "Nothing to compile. Sources did not change!")
-        (System/exit 0))
-    (binding [HL_NO_DOCKER? local]
-      (docker-run (str "clojure -X:uberjar :aliases '" (str [CLJ_ALIAS_KEY]) "' :aot '[\"" (str ENTRYPOINT) "\"]' " ":jvm-opts '[\"-Dclojure.compiler.direct-linking=true\", \"-Dclojure.spec.skip-macros=true\"]' :jar " OUTPUT_JAR_PATH " :main-class " (str ENTRYPOINT)))
+      (hpr "Nothing to compile. Sources did not change!")
+      (System/exit 0))
+    (binding [HL_NO_DOCKER? true]
+      (when-not COMPILE_CMD
+        (hpr (pre "Neither HL_COMPILE_CMD nor :build:compile-cmd has been specified!"))
+        (System/exit 1))
+      (docker-run COMPILE_CMD)
+      (stat-file ".holy-lambda/build/output.jar")
       (hpr "Uberjar artifact of the project is available at" (accent ".holy-lambda/build/output.jar")))))
-
-(defn mvn-local-test
-  [file]
-  (if (not= (:mvn/local-repo (edn/read-string (slurp (io/file file)))) ".holy-lambda/.m2")
-    (hpr (pre "property") (accent ":mvn/local-repo") (pre "in file") (accent file) (pre "should be set to") (accent ".holy-lambda/.m2"))
-    (hpr (prs "property") (accent ":mvn/local-repo") (prs "in file") (accent file) (prs "is correct"))))
 
 (defn hl:doctor
   "     \033[0;31m>\033[0m Diagnoses common issues in the project"
@@ -623,7 +596,6 @@ set -e
       (hpr " Babashka tasks version:  " (accent TASKS_VERSION))
       (hpr " Babashka version:        " (accent (or (s/trim (shs-no-err "bb" "version")) "UNKNOWN")))
       (hpr " TTY:                     " (accent TTY?))
-      (hpr " Entrypoint:              " (accent ENTRYPOINT))
       (hpr "---------------------------------------\n"))
 
     (when-not (file-exists? AWS_DIR)
@@ -634,22 +606,9 @@ set -e
       (hpr (pre "File deps.edn does not exists!"))
       (exit-code-err!))
 
-    (if-not ENTRYPOINT
-      (hpr (pre ":runtime:entrypoint is required!"))
-      (hpr (prs ":runtime:entrypoint looks good")))
-
     (when (and NATIVE_DEPS_PATH
                (not (file-exists? NATIVE_DEPS_PATH)))
-      (hpr (prw ":runtime:native-deps folder does not exists") (accent ":native:executable") (prw "will not include any extra deps!")))
-
-    (mvn-local-test "deps.edn")
-    (mvn-local-test "bb.edn")
-
-    (if (file-exists? HOLY_LAMBDA_DEPS_PATH)
-      (hpr (prs "Syncing stack is not required"))
-      (do
-        (hpr (pre "Stack is not synced! Run:") (accent "hl:sync"))
-        (exit-code-err!)))
+      (hpr (prw ":backend:native-deps folder does not exists") (accent ":native:executable") (prw "will not include any extra deps!")))
 
     (if-let [cmds-not-found (seq (filter (comp not command-exists?) REQUIRED_COMMANDS))]
       (do
@@ -657,15 +616,21 @@ set -e
         (exit-code-err!))
       (do
         (hpr (prs "Required commands") (accent (str REQUIRED_COMMANDS)) (prs "installed!"))))
-    (System/exit @exit-code)))
+    (System/exit @exit-code))
 
-(defn hl:version
-  "     \033[0;31m>\033[0m Outputs holy-lambda babashka tasks version"
+  (defn hl:version
+    "     \033[0;31m>\033[0m Outputs holy-lambda babashka tasks version"
+    []
+    (print-task "hl:version")
+    (hpr (str (prs "Current tasks version is: ") (accent TASKS_VERSION)))
+    (when-not (= (s/trim (:body (curl/get REMOTE_TASKS))) TASKS_VERSION)
+      (hpr (prw "Local version of tasks does not match stable tasks version. Update tasks sha!")))))
+
+(defn hl:sync
+  "     \033[0;31m>\033[0m DEPRECATED! See HL changelog"
   []
-  (print-task "hl:version")
-  (hpr (str (prs "Current tasks version is: ") (accent TASKS_VERSION)))
-  (when-not (= (s/trim (:body (curl/get REMOTE_TASKS))) TASKS_VERSION)
-    (hpr (prw "Local version of tasks does not match stable tasks version. Update tasks sha!"))))
+  (print-task "hl:sync")
+  (hpr (pre "hl:sync is no-op. Clojure dependencies are now downloaded to $HOME/.m2 directory. Remove :mvn/local-repo from the deps.edn. For downloading babashka pods and dependencies execute hl:babashka:sync!")))
 
 (defn hl:clean
   "     \033[0;31m>\033[0m Cleanes build artifacts"